+++ chuchu/textbox.py
@@ -8,8 +8,15 @@ from chuchu.widget import TextWidget
 class Textbox(TextWidget):
     _TK_CLASS = tk.Entry
 
-    def __init__(self, text: str = "", *, validator: Callable[[str], bool] | None = None, **kwargs: Any) -> None:
-        super().__init__(text=text, validator=validator, **kwargs)
+    def __init__(
+        self,
+        text: str = "",
+        *,
+        validator: Callable[[str], bool] | None = None,
+        onchange: Callable[[str], Any] | None = None,
+        **kwargs: Any
+    ) -> None:
+        super().__init__(text=text, validator=validator, value=text, onchange=onchange, **kwargs)
 
     def write(self, text: str) -> int:
         """Write the given text to the end of the textbox, returning the number of characters written."""